New modeling backend structure (issue #282).
Changes
-------

All backend related code is in the modbackend directory. The main controlling
file is modeling.py, it decides which backend to load depending on the models
availability and the evironment variable CLMM_MODELING_BACKEND value:

- 'ct', or if it is unset, chooses cluster_toolkit + astropy;
- 'nc' chooses NumCosmo backend;
- 'ccl' chooses CCL backend (incomplete);

Both NumCosmo and CCL implement the abstract class CLMModeling
(clmm_modeling.py). This class holds the documentation for all implementations
(TODO) as well as all other details like units, little-h, etc. We also need to agree
on the names and arguments of the methods.

In addition I added a thin layer on top of the OO implementation called
func_layer.py. It mimics the current implementation based on cluster_toolkit.
Thus, all the other implementations will have the same behavior.

The pytest already works with all backends (CCL is currently removed from the
tests), it uses a module scoped fixture to control which backend to use.
Any code can try a different backend by calling
`os.environ['CLMM_MODELING_BACKEND'] = 'nc'`
**before** importing clmm.

In my opinion the ideal is to migrate to a OO interface (like the one in
NumCosmo and CCL), this can be done by instantiating the NumCosmo or
CCL implementation of CLMModel class instead of using the functional layer.

Other points:
-------------

- Increased a bit the tolerance value required in the distance comparison
  (1.0e-8 =>1.2e-8), CCL does not match with NumCosmo and Astropy at
  1.0e-8 level.
- The cosmology is still defined by a simple dictionary. We have to decide
  how to represent cosmology in a backend independent way.
- CCL and NumCosmo/cluster_toolkit do not agree already in the 3d
  density comparison. It does not differ only by an overall factor, I tried
  fixing that but the difference increases with R.
- Some functionalities are still not present on CCL.
- In CCL the mass-concentration relation is part of the implementation
  of the objects necessary to CLMM. In NumCosmo, Colossus and
  cluster_toolkit, that can be included by setting the concentration
  parameter using the appropriated c(M) relation (usually, these are
  very simple fitting functions). In any case, CCL have a
  `ConcentrationConstant` object that can be used to mimic the other
  codes behavior.
- For the moment being I wrote everything such that it agrees with the
  current results (therefore with cluster_toolkit). I think now it is a good
  moment to decide if the current input, units and little-h choices are
  OK.

Possible bugs
-------------

- In modbackend/cluster_toolkit.py, the implementation of
  angular_diameter_dist_a1a2 uses astropy to computes the distance
  and consequently to_value(units.pc) to compute it in unit of pc.
  However, there is a variable in constants.py (PC_TO_METER) which
  also contain the value of pc in meters. If these two disagree we can
  have inconsistencies. It may be safe, depends on how astropy
  computes the distances.
- In CCL massdef depends on concentration and concentration
  depends on massdef, am I missing something here? To make it work
  I just put by hand one on another after both were created.